+++ src/lib.rs
@@ -4,7 +4,7 @@ pub enum ModelError {
     /// Indicates that the model has not been fitted yet
     NotFitted,
     /// indicates the input data provided does not meet the expected format, type, or validation rules.
-    InputValidationError(&'static str),
+    InputValidationError(String),
     /// indicates that there is something wrong with the tree
     TreeError(&'static str),
 }